requirements.txt
python-dotenv
fastapi
uvicorn
requests

